requirements.txt
asgiref==3.5.2
astroid==3.0.1
attrs==22.2.0
beautifulsoup4==4.11.2
black==23.11.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.8.0rc2
dill==0.3.7
Django==3.2.16
django-bootstrap5==22.2
django-debug-toolbar==4.2.0
django-filter==23.3
django-stubs==4.2.6
django-stubs-ext==4.2.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Faker==12.0.1
flake8==6.0.0
flake8-docstrings==1.7.0
flake8-isort==6.0.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mixer==7.2.2
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.25.2
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.2
pep8-naming==0.13.3
php==1.2.1
Pillow==9.3.0
platformdirs==3.11.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
PyJWT==1.7.1
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-subtests==0.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.7
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soupsieve==2.5
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.12.2
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
yapf==0.32.0
